update 1.0
added checkuserexists and checkpassexists method to check if user or
pass exists in our database. updated the getconnection method to connect
to our new server. also added a verifyuser method and verifypass to pass
a true value for user and pass when logging in to make sure user and
pass exists in the same row.